Rename key files.
